fix: strictly validate regular Lagoon groups from Keycloak
Previously if a Keycloak group was not found to have the attributes of a
"project group", it was assumed to be a "regular group" in Lagoon.

It turns out that this is a bad assumption due to (among other things),
https://github.com/uselagoon/lagoon/pull/3426.

Therefore validate the "regular" Lagoon groups rather than making any
assumptions.