fix: use the opensearch pagination API correctly
Previously, opensearch-sync used the last value of the order-by field
(last_updated) in the search_after field of queries for subsequent
search result pages in the Opensearch API. This is flat-out wrong and
doesn't work!

Instead, the sort field of the last result must be passed in the
search_after field of subsequent queries to get the "next page" of
results. This API behaviour is described in the Paginate results
documentation.

https://docs.opensearch.org/2.19/search-plugins/searching-data/paginate/

This change also adds the ability to specify the size and search_after
fields in the dump-index-patterns debug command, to facilitate debugging
of index-pattern paging behaviour.