fix: reject partial search results for index patterns
If Opensearch is in a bad state it can return partial search results.

This is not useful for the opensearch sync tool because it will try to
create a bunch of index patterns that already exist, and throw a bunch
of errors.

Better to fail and just retry again on the next sync period.