fix: handle indexpatterns with identical sort values split across pages
When requesting index patterns, the updatedAt field is used for sorting.

Previously, when paging through index pattern search results, the sort
value of the last result of the current page was used in the searchAfter
field of the request for the next page. This works if all results have
unique sort values. However, index pattern results with the same name
can have identical sort values across multiple index migration versions.

Consider a run of results with identical sort values which are split
across two pages of results. Setting searchAfter equal to the sort value
of the last result of the first page, when requesting the second page,
has the effect of skipping the result values included in the run which
would otherwise appear on the second page.

This commit changes the logic so that the sort value of the last result
on the current page is decremented before being used in the searchAfter
field of the query for the next page. This ensures that results are no
longer skipped, but has the side effect of returning some results twice
when they are in a run of identical sort values split across pages. To
account for this, logic has been added to deduplicate results by index
pattern ID.